modules/module3_reasoning/requirements.txt
langchain==0.2.11
langchain-openai==0.1.17
openai==1.109.1
python-dotenv==1.0.0